test(*): add configurations to clear the database and export the app"
