+++ server/jest.config.js
@@ -1,4 +1,11 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  moduleFileExtensions: ['js', 'json', 'ts'],
+  rootDir: '',
+  testRegex: '.spec.ts$',
+  transform: {
+    '^.+\\.(t|j)s$': 'ts-jest',
+  },
+  setupFilesAfterEnv: ['./setupTests.ts'],
 };
