+++ server/package.json
@@ -10,8 +10,7 @@
     "watch:data": "if [ ! -f \"./dist/src/data/products.json\" ]; then npm run copy:json; fi && onchange 'src/data/*.json' -- npm run copy:json",
     "build": "env-cmd -f .env.${env} tsc && npm run copy:json",
     "deploy": "env-cmd -f .env.${env} npm run build && vercel --prod",
-    "test": "env-cmd -f .env.${env} jest --verbose --coverage ./tests/**/*.spec.ts",
-    "prettier": "prettier --write 'src/**/*.{ts,js}'",
+    "test": "env-cmd -f .env.${env} jest --verbose --coverage  --config ./jest.config.js --detectOpenHandles",
     "lint": "eslint '**/*.ts'"
   },
   "keywords": [],
@@ -22,8 +21,8 @@
     "cors": "^2.8.5",
     "env-cmd": "^10.1.0",
     "express": "^4.18.2",
-    "nodemailer": "^6.9.1",
     "express-async-errors": "^3.1.1",
+    "nodemailer": "^6.9.1",
     "pino": "^8.10.0",
     "typescript": "^4.9.5",
     "uuid": "^9.0.0"
@@ -31,17 +30,21 @@
   "devDependencies": {
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.5.0",
     "@types/nodemailer": "^6.4.7",
     "@types/prettier": "^2.7.2",
+    "@types/supertest": "^2.0.12",
     "@types/uuid": "^9.0.0",
     "@typescript-eslint/parser": "^5.51.0",
     "concurrently": "^7.6.0",
     "copyfiles": "^2.4.1",
     "eslint-plugin-no-unused-expressions": "^0.1.0",
+    "jest": "^29.5.0",
     "nodemon": "^2.0.20",
     "onchange": "^7.1.0",
     "prettier": "^2.8.4",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.0",
     "tsc-watch": "^6.0.0"
   }
 }
